requirements.txt
google-adk>=1.18.0
google-genai>=0.2.0
google-generativeai>=0.3.0
litellm>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pytest>=7.4.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
colorama>=0.4.6
